Division into translation units
